+++ packages/hot-updater/package.json
@@ -53,7 +53,6 @@
   },
   "dependencies": {
     "@clack/prompts": "^0.9.0",
-    "typescript": "5.8.2",
     "@hot-updater/console": "0.12.7",
     "@hot-updater/core": "0.12.7",
     "@hot-updater/plugin-core": "0.12.7",
@@ -62,15 +61,15 @@
     "cosmiconfig-typescript-loader": "^5.0.0"
   },
   "devDependencies": {
-    "@babel/parser": "^7.26.9",
-    "@babel/generator": "^7.26.9",
-    "@types/babel__generator": "^7.6.8",
-    "@babel/core": "^7.26.0",
-    "@babel/traverse": "^7.25.9",
-    "@babel/types": "^7.26.0",
+    "@babel/core": "7.26.0",
+    "@babel/traverse": "7.25.9",
+    "@babel/types": "7.26.0",
+    "@babel/parser": "7.26.9",
+    "@babel/generator": "7.26.9",
+    "@types/babel__generator": "7.6.8",
     "@hono/node-server": "^1.13.4",
-    "@types/babel__core": "^7.20.5",
-    "@types/babel__traverse": "^7.20.6",
+    "@types/babel__core": "7.20.5",
+    "@types/babel__traverse": "7.20.6",
     "@types/connect": "^3.4.38",
     "@types/cosmiconfig": "^6.0.0",
     "@types/node": "^22.9.0",
