+++ packages/hot-updater/src/utils/printBanner.ts
@@ -1,0 +1,6 @@
+import { version } from "@/packageJson";
+import { printBanner as _printBanner } from "@hot-updater/plugin-core";
+
+export const printBanner = () => {
+  _printBanner(version);
+};
