+++ plugins/aws/tsup.config.ts
@@ -2,7 +2,7 @@ import { defineConfig } from "tsup";
 
 export default defineConfig([
   {
-    entry: ["src/index.ts", "sdk/index.ts", "migrations/index.ts"],
+    entry: ["src/index.ts"],
     format: ["esm", "cjs"],
     outDir: "dist",
     dts: true,
@@ -14,5 +14,15 @@ export default defineConfig([
     entry: ["lambda/index.ts"],
     format: ["cjs"],
     outDir: "dist/lambda",
+  },
+  {
+    entry: ["iac/index.ts"],
+    format: ["esm", "cjs"],
+    dts: true,
+    outDir: "dist/iac",
+    external: ["@hot-updater/aws"],
+    banner: {
+      js: `import { createRequire } from 'module'; const require = createRequire(import.meta.url);`,
+    },
   },
 ]);
