+++ plugins/plugin-core/src/index.ts
@@ -4,3 +4,9 @@ export * from "./loadConfig";
 export * from "./types";
 export * from "./copyDirToTmp";
 export * from "./createDatabasePlugin";
+export * from "./banner";
+export * from "./transformTemplate";
+export * from "./transformEnv";
+export * from "./transformTsEnv";
+export * from "./makeEnv";
+export * from "./createZip";
