+++ plugins/plugin-core/src/transformEnv.spec.ts
