+++ plugins/plugin-core/src/transformEnv.ts
