+++ plugins/plugin-core/src/transformTemplate.ts
