+++ plugins/plugin-core/src/transformTsEnv.spec.ts
