+++ plugins/plugin-core/src/transformTsEnv.ts
